HID anti-cheat: 0xFF00->0x0001 usage page, template-matching crosshair, debug aim + overlay box, crosshair latency metric

+++ src/MouseController.cpp
@@ -152,9 +152,9 @@ bool MouseController::processHIDReport(std::vector<uint8_t> &report) {
     }
 
     DWORD bytesWritten = 0;
-    BOOL res = WriteFile(hidDevice, report.data(), report.size(), &bytesWritten, NULL);
+    BOOL res = WriteFile(hidDevice, report.data(), (DWORD)report.size(), &bytesWritten, NULL);
     if (res && bytesWritten == report.size()) {
-        hidWarningLogged = false; // back to healthy — re-arm the one-shot
+        hidWarningLogged = false;
         return true;
     }
